+++ packages/server/src/utils/odata.test.ts
@@ -1,125 +1,300 @@
-import { Op, Order, WhereOptions } from 'sequelize';
+import { and, col, fn, json, Op, or, Order, where, WhereOptions } from 'sequelize';
 
 import { odataFilterToSequelize, odataOrderbyToSequelize } from './odata';
 
 describe('odataFilterToSequelize', () => {
   const cases: { [key: string]: WhereOptions } = {
-    // Simple equal to
-    'foo eq true': { foo: { [Op.eq]: true } },
-    'foo eq 1999-12-31': { foo: { [Op.eq]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo eq 12': { foo: { [Op.eq]: 12 } },
-    "foo eq 'bar'": { foo: { [Op.eq]: 'bar' } },
+    //
+    // 5.1.1.1 Logical Operators
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_LogicalOperators
+    //
 
-    // Simple not equal to
-    'foo ne true': { foo: { [Op.ne]: true } },
-    'foo ne 1999-12-31': { foo: { [Op.ne]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo ne 12': { foo: { [Op.ne]: 12 } },
-    "foo ne 'bar'": { foo: { [Op.ne]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360958
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo eq true': where(col('Model.foo'), '=', true),
+    'foo eq 01234567-89ab-cdef-0123-456789abcdef': where(
+      col('Model.foo'),
+      '=',
+      '01234567-89ab-cdef-0123-456789abcdef',
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo eq 1999-12-31': where(col('Model.foo'), '=', new Date('1999-12-31T00:00:00Z')),
+    'foo eq 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '=',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo eq 12': where(col('Model.foo'), '=', 12),
+    "foo eq 'bar'": where(col('Model.foo'), '=', 'bar'),
 
-    // Simple lesser than
-    'foo lt 1999-12-31': { foo: { [Op.lt]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo lt 12': { foo: { [Op.lt]: 12 } },
-    "foo lt 'bar'": { foo: { [Op.lt]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360959
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo ne true': where(col('Model.foo'), '!=', true),
+    'foo ne 01234567-89ab-cdef-0123-456789abcdef': where(
+      col('Model.foo'),
+      '!=',
+      '01234567-89ab-cdef-0123-456789abcdef',
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo ne 1999-12-31': where(col('Model.foo'), '!=', new Date('1999-12-31T00:00:00Z')),
+    'foo ne 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '!=',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo ne 12': where(col('Model.foo'), '!=', 12),
+    "foo ne 'bar'": where(col('Model.foo'), '!=', 'bar'),
 
-    // Simple lesser than or equal to
-    'foo le 1999-12-31': { foo: { [Op.lte]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo le 12': { foo: { [Op.lte]: 12 } },
-    "foo le 'bar'": { foo: { [Op.lte]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360960
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo gt 1999-12-31': where(col('Model.foo'), '>', new Date('1999-12-31T00:00:00Z')),
+    'foo gt 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '>',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo gt 12': where(col('Model.foo'), '>', 12),
+    "foo gt 'bar'": where(col('Model.foo'), '>', 'bar'),
 
-    // Simple greater than
-    'foo gt 1999-12-31': { foo: { [Op.gt]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo gt 12': { foo: { [Op.gt]: 12 } },
-    "foo gt 'bar'": { foo: { [Op.gt]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360961
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo ge 1999-12-31': where(col('Model.foo'), '>=', new Date('1999-12-31T00:00:00Z')),
+    'foo ge 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '>=',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo ge 12': where(col('Model.foo'), '>=', 12),
+    "foo ge 'bar'": where(col('Model.foo'), '>=', 'bar'),
 
-    // Simple greater than or equal to
-    'foo ge 1999-12-31': { foo: { [Op.gte]: new Date('1999-12-31T00:00:00.000Z') } },
-    'foo ge 12': { foo: { [Op.gte]: 12 } },
-    "foo ge 'bar'": { foo: { [Op.gte]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360962
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo lt 1999-12-31': where(col('Model.foo'), '<', new Date('1999-12-31T00:00:00Z')),
+    'foo lt 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '<',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo lt 12': where(col('Model.foo'), '<', 12),
+    "foo lt 'bar'": where(col('Model.foo'), '<', 'bar'),
 
-    // Functions
-    "startswith(foo, 'bar')": { foo: { [Op.startsWith]: 'bar' } },
-    "endswith(foo, 'bar')": { foo: { [Op.endsWith]: 'bar' } },
-    "substringof(foo, 'bar')": { foo: { [Op.substring]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360963
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo le 1999-12-31': where(col('Model.foo'), '<=', new Date('1999-12-31T00:00:00Z')),
+    'foo le 1999-12-31T12:00:00+02:00': where(
+      col('Model.foo'),
+      '<=',
+      // @ts-expect-error This is a bug in the Sequelize types.
+      new Date('1999-12-31T10:00:00Z'),
+    ),
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo le 12': where(col('Model.foo'), '<=', 12),
+    "foo le 'bar'": where(col('Model.foo'), '<=', 'bar'),
 
-    // Nested equal to
-    'foo/bar/baz eq true': { 'foo.bar.baz': { [Op.eq]: true } },
-    'foo/bar/baz eq 1999-12-31': { 'foo.bar.baz': { [Op.eq]: '1999-12-31' } },
-    'foo/bar/baz eq 12': { 'foo.bar.baz': { [Op.eq]: 12 } },
-    "foo/bar/baz eq 'bar'": { 'foo.bar.baz': { [Op.eq]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360964
+    'foo eq 12 and bar eq 14': and(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+    ),
+    'foo eq 12 and bar eq 14 and baz eq 8': and(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.baz'), '=', 8),
+    ),
+    'foo eq 12 and (bar eq 14 and baz eq 8)': and(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.baz'), '=', 8),
+    ),
 
-    // Nested not equal to
-    'foo/bar/baz ne true': { 'foo.bar.baz': { [Op.ne]: true } },
-    'foo/bar/baz ne 1999-12-31': { 'foo.bar.baz': { [Op.ne]: '1999-12-31' } },
-    'foo/bar/baz ne 12': { 'foo.bar.baz': { [Op.ne]: 12 } },
-    "foo/bar/baz ne 'bar'": { 'foo.bar.baz': { [Op.ne]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#_Toc31360965
+    'foo eq 12 or bar eq 14': or(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+    ),
+    'foo eq 12 or bar eq 14 or baz eq 8': or(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.baz'), '=', 8),
+    ),
+    'foo eq 12 or (bar eq 14 or baz eq 8)': or(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.bar'), '=', 14),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.baz'), '=', 8),
+    ),
 
-    // Nested lesser than
-    'foo/bar/baz lt 1999-12-31': { 'foo.bar.baz': { [Op.lt]: '1999-12-31' } },
-    'foo/bar/baz lt 12': { 'foo.bar.baz': { [Op.lt]: 12 } },
-    "foo/bar/baz lt 'bar'": { 'foo.bar.baz': { [Op.lt]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Not
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'not foo eq 12': { [Op.not]: where(col('Model.foo'), '=', 12) },
 
-    // Nested lesser than or equal to
-    'foo/bar/baz le 1999-12-31': { 'foo.bar.baz': { [Op.lte]: '1999-12-31' } },
-    'foo/bar/baz le 12': { 'foo.bar.baz': { [Op.lte]: 12 } },
-    "foo/bar/baz le 'bar'": { 'foo.bar.baz': { [Op.lte]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Has
+    // XXX
 
-    // Nested greater than
-    'foo/bar/baz gt 1999-12-31': { 'foo.bar.baz': { [Op.gt]: '1999-12-31' } },
-    'foo/bar/baz gt 12': { 'foo.bar.baz': { [Op.gt]: 12 } },
-    "foo/bar/baz gt 'bar'": { 'foo.bar.baz': { [Op.gt]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_In
+    // XXX
 
-    // Nested greater than or equal to
-    'foo/bar/baz ge 1999-12-31': { 'foo.bar.baz': { [Op.gte]: '1999-12-31' } },
-    'foo/bar/baz ge 12': { 'foo.bar.baz': { [Op.gte]: 12 } },
-    "foo/bar/baz ge 'bar'": { 'foo.bar.baz': { [Op.gte]: 'bar' } },
+    //
+    // 5.1.1.2 Arithmetic Operators
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_ArithmeticOperators
+    //
 
-    // Simple Functions
-    "startswith(foo/bar/baz, 'bar')": { 'foo.bar.baz': { [Op.startsWith]: 'bar' } },
-    "endswith(foo/bar/baz, 'bar')": { 'foo.bar.baz': { [Op.endsWith]: 'bar' } },
-    "substringof(foo/bar/baz, 'bar')": { 'foo.bar.baz': { [Op.substring]: 'bar' } },
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Addition
+    // @ts-expect-error This is a bug in the Sequelize types
+    'foo add 1 eq 3': where(where(col('Model.foo'), '+', 1), '=', 3),
 
-    // Logical statements
-    'foo eq 12 and bar eq 14': {
-      [Op.and]: [{ foo: { [Op.eq]: 12 } }, { bar: { [Op.eq]: 14 } }],
-    },
-    'foo eq 12 and bar eq 14 and baz eq 8': {
-      [Op.and]: [
-        { foo: { [Op.eq]: 12 } },
-        { [Op.and]: [{ bar: { [Op.eq]: 14 } }, { baz: { [Op.eq]: 8 } }] },
-      ],
-    },
-    'foo eq 12 and (bar eq 14 and baz eq 8)': {
-      [Op.and]: [
-        { foo: { [Op.eq]: 12 } },
-        { [Op.and]: [{ bar: { [Op.eq]: 14 } }, { baz: { [Op.eq]: 8 } }] },
-      ],
-    },
-    'foo eq 12 or bar eq 14': {
-      [Op.or]: [{ foo: { [Op.eq]: 12 } }, { bar: { [Op.eq]: 14 } }],
-    },
-    'foo eq 12 or bar eq 14 or baz eq 8': {
-      [Op.or]: [
-        { foo: { [Op.eq]: 12 } },
-        { [Op.or]: [{ bar: { [Op.eq]: 14 } }, { baz: { [Op.eq]: 8 } }] },
-      ],
-    },
-    'foo eq 12 or (bar eq 14 or baz eq 8)': {
-      [Op.or]: [
-        { foo: { [Op.eq]: 12 } },
-        { [Op.or]: [{ bar: { [Op.eq]: 14 } }, { baz: { [Op.eq]: 8 } }] },
-      ],
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Subtraction
+    // @ts-expect-error This is a bug in the Sequelize types
+    'foo sub 1 eq 3': where(where(col('Model.foo'), '-', 1), '=', 3),
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Negation
+    // XXX
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Multiplication
+    // @ts-expect-error This is a bug in the Sequelize types
+    'foo mul 1 eq 3': where(where(col('Model.foo'), '*', 1), '=', 3),
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Divisionmul
+    // @ts-expect-error This is a bug in the Sequelize types
+    'foo div 1 eq 3': where(where(col('Model.foo'), '/', 1), '=', 3),
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_Modulo
+    // @ts-expect-error This is a bug in the Sequelize types
+    'foo mod 1 eq 3': where(where(col('Model.foo'), '%', 1), '=', 3),
+
+    //
+    // String and Collection Functions
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_StringandCollectionFunctions
+    //
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_concat
+    "concat(foo, 'r') eq 'bar'": where(fn('concat', col('Model.foo'), 'r'), '=', 'bar'),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_contains
+    "contains(foo, 'bar')": where(col('Model.foo'), { [Op.substring]: 'bar' }),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_endswith
+    "endswith(foo, 'bar')": where(col('Model.foo'), { [Op.endsWith]: 'bar' }),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_indexof
+    // @ts-expect-error This is a bug in the Sequelize types
+    "indexof(foo, 'bar') eq 3": where(fn('strpos', col('Model.foo'), 'bar'), '=', 3),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_length
+    // @ts-expect-error This is a bug in the Sequelize types
+    'length(foo) eq 42': where(fn('length', col('Model.foo')), '=', 42),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_startswith
+    "startswith(foo, 'bar')": where(col('Model.foo'), { [Op.startsWith]: 'bar' }),
+    // XXX implement for collections
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_substring
+    "substring(foo, 1, 2) eq 'b'": where(fn('substring', col('Model.foo'), 1, 2), '=', 'b'),
+    // XXX implement for collections
+
+    //
+    // Collection Functions
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_CollectionFunctions
+    //
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_hassubset
+    // XXX
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_hassubsequence
+    // XXX
+
+    //
+    // String Functions
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_StringFunctions
+    //
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_matchesPattern
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_tolower
+    "tolower(foo) eq 'bar'": where(fn('lower', col('Model.foo')), '=', 'bar'),
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_toupper
+    "toupper(foo) eq 'bar'": where(fn('upper', col('Model.foo')), '=', 'bar'),
+
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_trim
+    "trim(foo) eq 'bar'": where(fn('trim', col('Model.foo')), '=', 'bar'),
+
+    //
+    // Date and Time Functions
+    // https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_DateandTimeFunctions
+    //
+
+    // Nested properties
+    // @ts-expect-error This is a bug in the Sequelize types.
+    'foo/bar/baz eq 42': where(json('foo.bar.baz'), '=', 42),
+
+    // Nested functions
+    "contains(tolower(foo), 'bar')": where(fn('lower', col('Model.foo')), {
+      [Op.substring]: 'bar',
+    }),
+
+    // Combine boolean logical expressions
+    'foo eq 12 or (bar eq 14 and baz eq 8)': or(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      and(where(col('Model.bar'), '=', 14), where(col('Model.baz'), '=', 8)),
+    ),
+    'foo eq 12 and (bar eq 14 or baz eq 8)': and(
+      // @ts-expect-error This is a bug in the Sequelize types.
+      where(col('Model.foo'), '=', 12),
+      // @ts-expect-error This is a bug in the Sequelize types.
+      or(where(col('Model.bar'), '=', 14), where(col('Model.baz'), '=', 8)),
+    ),
+    'not (foo eq 12 and bar eq 14)': {
+      [Op.not]: and(
+        // @ts-expect-error This is a bug in the Sequelize types.
+        where(col('Model.foo'), '=', 12),
+        // @ts-expect-error This is a bug in the Sequelize types.
+        where(col('Model.bar'), '=', 14),
+      ),
     },
   };
 
   it.each(Object.entries(cases))('%s', (filter, expected) => {
-    const result = odataFilterToSequelize(filter);
+    const result = odataFilterToSequelize(filter, { tableName: 'Model' });
     expect(result).toStrictEqual(expected);
   });
 
   const empty = ['', null, undefined];
 
   it.each(empty)('%p', (filter) => {
-    const result = odataFilterToSequelize(filter);
+    const result = odataFilterToSequelize(filter, { tableName: 'Model' });
     expect(result).toStrictEqual({});
   });
 });