+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.15.1] - 2020-10-05
+
+### Added
+
+- Block(`form`): Implemented `required` requirement for `date-time` fields.
+
+### Fixed
+
+- Block(`filter`): Fixed layout for highlighted `date-range` fields.
+- Server: Added support for more numeric OData types.
+
+## [0.15.0] - 2020-10-02
+
+### Added
+
+- Block(`filter`): Made the search for string fields case insensitive.
+- Block(`form`): Added support for length requirements for repeated object fields.
+- Block(`table`): Add support for `dropdown` fields.
+
+### Fixed
+
+- Block(`form`): Fixed the accept property for file fields.
+
 ## [0.14.2] - 2020-09-25
 
 ### Fixed